requirements.txt
streamlit
pandas
numpy
torch
requests
Pillow
transformers
open_clip_torch
scikit-learn
matplotlib